fix: Move only files, not folders

+++ classifier/classifier.py
@@ -15,13 +15,15 @@ Documents 	-	https://en.wikipedia.org/wiki/List_of_Microsoft_Office_filename_ext
 
 
 def moveto(file, from_folder, to_folder):
-    if not os.path.exists(to_folder):
-        os.makedirs(to_folder)
-
     from_file = os.path.join(from_folder, file)
     to_file = os.path.join(to_folder, file)
 
-    os.rename(from_file, to_file)
+    #to move only files, not folders
+    if(os.path.isfile(from_file)):
+        if not os.path.exists(to_folder):
+            os.makedirs(to_folder)
+            
+        os.rename(from_file, to_file)
 
 
 def classify(formats, output):
